+++ zed_components/src/zed_camera/src/zed_camera_component.cpp
@@ -2617,7 +2617,6 @@ bool ZedCamera::startCamera()
   // ----> Initialize Diagnostic statistics
   mElabPeriodMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
   mGrabPeriodMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
-  mGrabFuncMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
   mVideoDepthPeriodMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
   mVideoDepthElabMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
   mPcPeriodMean_sec = std::make_unique<sl_tools::WinAvg>(mCamGrabFrameRate);
@@ -3382,7 +3381,6 @@ void ZedCamera::threadFunc_zedGrab()
   RCLCPP_DEBUG(get_logger(), "Grab thread started");
 
   mFrameCount = 0;
-  mZedDataReady = false;
 
   // ----> Grab Runtime parameters
   mRunParams.sensing_mode = static_cast<sl::SENSING_MODE>(mDepthSensingMode);
@@ -3447,10 +3445,6 @@ void ZedCamera::threadFunc_zedGrab()
     }
     // ----> Check for Object Detection requirement
 
-    // ----> Wait for RGB/Depth retrieving before grabbing
-    std::unique_lock<std::mutex> datalock(mCamDataMutex);
-    // <---- Wait for RGB/Depth retrieving before grabbing
-
     // ----> Grab freq calculation
     static sl_tools::StopWatch grabFreqTimer;
 
@@ -3464,27 +3458,19 @@ void ZedCamera::threadFunc_zedGrab()
     // <---- Grab freq calculation
 
     if (!mSvoPause) {
-
-      sl_tools::StopWatch grabTimer;
+      // Start processing timer for diagnostic
+      grabElabTimer.tic();
 
       // ZED grab
       mGrabStatus = mZed.grab(mRunParams);
 
-      // Diagnostic statistics update
-      double grab_sec = grabTimer.toc();
-      double mean_grab_sec = mGrabFuncMean_sec->addValue(grab_sec);
-
-      if (mDebugMode) {
-        RCLCPP_DEBUG_STREAM(
-          get_logger(), "*** sl::Camera::grab time: " << grab_sec << " sec");
+      // ----> Grab errors?
+      // Note: disconnection are automatically handled by the ZED SDK
+      if (mGrabStatus != sl::ERROR_CODE::SUCCESS) {
+        RCLCPP_ERROR_STREAM(get_logger(), "Camera error: " << sl::toString(mGrabStatus).c_str());
+        break;  // TODO(walter) verify what to do in case of grab errors
       }
-
-      // Start processing timer for diagnostic
-      grabElabTimer.tic();
-
-      // ----> Unlock RGB/Depth thread
-      mZedDataReady = true;
-      // <---- Unlock RGB/Depth thread
+      // <---- Grab errors?
 
       // ----> Check SVO status
       if (mSvoMode && mGrabStatus == sl::ERROR_CODE::END_OF_SVOFILE_REACHED) {
@@ -3500,14 +3486,6 @@ void ZedCamera::threadFunc_zedGrab()
         }
       }
       // <---- Check SVO status
-
-      // ----> Grab errors?
-      // Note: disconnection are automatically handled by the ZED SDK
-      if (mGrabStatus != sl::ERROR_CODE::SUCCESS) {
-        RCLCPP_ERROR_STREAM(get_logger(), "Camera error: " << sl::toString(mGrabStatus).c_str());
-        break;  // TODO(walter) verify what to do in case of grab errors
-      }
-      // <---- Grab errors?
 
       mFrameCount++;
 
@@ -3533,6 +3511,28 @@ void ZedCamera::threadFunc_zedGrab()
       // <---- Check recording status
     }
 
+    // ----> Retrieve Image/Depth data if someone has subscribed to
+    // Retrieve data if there are subscriber to topics
+    if (areVideoDepthSubscribed()) {
+      // Run the point cloud conversion asynchronously to avoid slowing down
+      // all the program
+      // Retrieve raw pointCloud data if latest Pointcloud is ready
+      std::unique_lock<std::mutex> vd_lock(mVideoDepthMutex, std::defer_lock);
+
+      if (vd_lock.try_lock()) {
+        RCLCPP_DEBUG(get_logger(), "Retrieving video/depth data");
+        retrieveVideoDepth();
+
+        // Signal Pointcloud thread that a new pointcloud is ready
+        mVideoDepthDataReadyCondVar.notify_one();
+        mVideoDepthDataReady = true;
+        mVdPublishing = true;
+      }
+    } else {
+      mVdPublishing = false;
+    }
+    // <---- Retrieve Image/Depth data if someone has subscribed to
+
     if (!mDepthDisabled) {
       if (mPosTrackingStarted) {
         if (!mSvoPause) {
@@ -3549,7 +3549,7 @@ void ZedCamera::threadFunc_zedGrab()
       }
     }
 
-    // Publish the point cloud if someone has subscribed to
+    // ----> Retrieve the point cloud if someone has subscribed to
     if (!mDepthDisabled) {
       size_t cloudSubnumber = 0;
       try {
@@ -3565,9 +3565,9 @@ void ZedCamera::threadFunc_zedGrab()
         // Run the point cloud conversion asynchronously to avoid slowing down
         // all the program
         // Retrieve raw pointCloud data if latest Pointcloud is ready
-        std::unique_lock<std::mutex> lock(mPcMutex, std::defer_lock);
+        std::unique_lock<std::mutex> pc_lock(mPcMutex, std::defer_lock);
 
-        if (lock.try_lock()) {
+        if (pc_lock.try_lock()) {
           RCLCPP_DEBUG(get_logger(), "Retrieving point cloud");
           mZed.retrieveMeasure(mMatCloud, sl::MEASURE::XYZBGRA, sl::MEM::CPU, mMatResol);
 
@@ -3580,6 +3580,7 @@ void ZedCamera::threadFunc_zedGrab()
         mPcPublishing = false;
       }
     }
+    // <---- Retrieve the point cloud if someone has subscribed to
 
     if (!mDepthDisabled) {
       mObjDetMutex.lock();
@@ -4060,9 +4061,10 @@ void ZedCamera::publishPoseTF(rclcpp::Time t)
 
 void ZedCamera::threadFunc_pointcloudElab()
 {
-  std::unique_lock<std::mutex> lock(mPcMutex);
-
   RCLCPP_DEBUG(get_logger(), "Point Cloud thread started");
+  mPcDataReady = false;
+
+  std::unique_lock<std::mutex> lock(mPcMutex);
 
   while (1) {
     if (!rclcpp::ok()) {
@@ -4124,52 +4126,73 @@ void ZedCamera::threadFunc_pointcloudElab()
 
 void ZedCamera::threadFunc_pubVideoDepth()
 {
-  RCLCPP_DEBUG_ONCE(get_logger(), "Video Depth thread started");
+  RCLCPP_DEBUG(get_logger(), "Video Depth thread started");
+  mVideoDepthDataReady = false;
+
+  std::unique_lock<std::mutex> lock(mVideoDepthMutex);
 
   while (1) {
     if (!rclcpp::ok()) {
-      RCLCPP_DEBUG(get_logger(), "Ctrl+C received: stopping rgb/depth pub thread");
-      mThreadStop = true;
-      break;
-    }
-    if (mThreadStop) {
-      RCLCPP_DEBUG(get_logger(), "threadFunc_pubVideoDepth (2): thread stopped");
+      RCLCPP_DEBUG(get_logger(), "Ctrl+C received: stopping video depth thread");
       break;
     }
 
-    if (!mZed.isOpened()) {
-      RCLCPP_DEBUG(get_logger(), "threadFunc_pubVideoDepth: the camera is not open");
-      continue;
+    // RCLCPP_DEBUG(get_logger(), "threadFunc_pubVideoDepth -> mVideoDepthDataReady value:
+    // %s", mVideoDepthDataReady ? "TRUE" : "FALSE");
+
+    while (!mVideoDepthDataReady) {  // loop to avoid spurious wakeups
+      if (
+        mVideoDepthDataReadyCondVar.wait_for(lock, std::chrono::milliseconds(500)) ==
+        std::cv_status::timeout)
+      {
+        // Check thread stopping
+        if (!rclcpp::ok()) {
+          RCLCPP_DEBUG(get_logger(), "Ctrl+C received: stopping video depth thread");
+          mThreadStop = true;
+          break;
+        }
+        if (mThreadStop) {
+          RCLCPP_DEBUG(get_logger(), "threadFunc_pubVideoDepth (2): video depth thread stopped");
+          break;
+        } else {
+          // RCLCPP_DEBUG(get_logger(), "threadFunc_pubVideoDepth -> WAIT FOR DATA");
+          continue;
+        }
+      }
+    }
+
+    if (mThreadStop) {
+      RCLCPP_DEBUG(get_logger(), "threadFunc_pubVideoDepth (1): video depth thread stopped");
+      break;
     }
 
     rclcpp::Time pub_ts;
-    mPublishingData = publishVideoDepth(pub_ts);
+    publishVideoDepth(pub_ts);
 
-    // RCLCPP_INFO_STREAM(get_logger(), "threadFunc_pubVideoDepth - pub_ts type:" <<
-    // pub_ts.get_clock_type());
-
-    if (mCamRealModel != sl::MODEL::ZED && mPublishingData && pub_ts != TIMEZERO_ROS) {
+    if (mCamRealModel != sl::MODEL::ZED && mVdPublishing && pub_ts != TIMEZERO_ROS) {
       if (mSensCameraSync || mSvoMode) {
         publishSensorsData(pub_ts);
       }
     }
 
     // ----> Check publishing frequency
-    double pub_period_usec = 1e6 / mPubFrameRate;
+    double vd_period_usec = 1e6 / mPubFrameRate;
 
-    static sl_tools::StopWatch pubFreqTimer;
-    double elapsed_usec = pubFreqTimer.toc() * 1e6;
+    static sl_tools::StopWatch vdPubFreqTimer;
+    double elapsed_usec = vdPubFreqTimer.toc() * 1e6;
 
-    if (elapsed_usec < pub_period_usec) {
-      rclcpp::sleep_for(
-        std::chrono::microseconds(static_cast<int>(pub_period_usec - elapsed_usec)));
+    if (elapsed_usec < vd_period_usec) {
+      rclcpp::sleep_for(std::chrono::microseconds(static_cast<int>(vd_period_usec - elapsed_usec)));
     }
 
-    pubFreqTimer.tic();
+    vdPubFreqTimer.tic();
     // <---- Check publishing frequency
+
+    mVideoDepthDataReady = false;
+    // RCLCPP_DEBUG(get_logger(), "threadFunc_pubVideoDepth -> mVideoDepthDataReady FALSE")
   }
 
-  RCLCPP_DEBUG(get_logger(), "Video/Depth thread finished");
+  RCLCPP_DEBUG(get_logger(), "Pointcloud thread finished");
 }
 
 void ZedCamera::threadFunc_pubSensorsData()
@@ -4228,204 +4251,202 @@ void ZedCamera::threadFunc_pubSensorsData()
   RCLCPP_DEBUG(get_logger(), "Sensors thread finished");
 }
 
-bool ZedCamera::publishVideoDepth(rclcpp::Time & out_pub_ts)
+bool ZedCamera::areVideoDepthSubscribed()
+{
+  mRgbSubnumber = 0;
+  mRgbRawSubnumber = 0;
+  mRgbGraySubnumber = 0;
+  mRgbGrayRawSubnumber = 0;
+  mLeftSubnumber = 0;
+  mLeftRawSubnumber = 0;
+  mLeftGraySubnumber = 0;
+  mLeftGrayRawSubnumber = 0;
+  mRightSubnumber = 0;
+  mRightRawSubnumber = 0;
+  mRightGraySubnumber = 0;
+  mRightGrayRawSubnumber = 0;
+  mStereoSubnumber = 0;
+  mStereoRawSubnumber = 0;
+  mDepthSubnumber = 0;
+  mConfMapSubnumber = 0;
+  mDisparitySubnumber = 0;
+  mDepthInfoSubnumber = 0;
+
+  try {
+    mRgbSubnumber = mPubRgb.getNumSubscribers();
+    mRgbRawSubnumber = mPubRawRgb.getNumSubscribers();
+    mRgbGraySubnumber = mPubRgbGray.getNumSubscribers();
+    mRgbGrayRawSubnumber = mPubRawRgbGray.getNumSubscribers();
+    mLeftSubnumber = mPubLeft.getNumSubscribers();
+    mLeftRawSubnumber = mPubRawLeft.getNumSubscribers();
+    mLeftGraySubnumber = mPubLeftGray.getNumSubscribers();
+    mLeftGrayRawSubnumber = mPubRawLeftGray.getNumSubscribers();
+    mRightSubnumber = mPubRight.getNumSubscribers();
+    mRightRawSubnumber = mPubRawRight.getNumSubscribers();
+    mRightGraySubnumber = mPubRightGray.getNumSubscribers();
+    mRightGrayRawSubnumber = mPubRawRightGray.getNumSubscribers();
+    mStereoSubnumber = mPubStereo.getNumSubscribers();
+    mStereoRawSubnumber = mPubRawStereo.getNumSubscribers();
+
+    if (!mDepthDisabled) {
+      mDepthSubnumber = mPubDepth.getNumSubscribers();
+      mDepthInfoSubnumber = count_subscribers(mPubDepthInfo->get_topic_name());
+      mConfMapSubnumber = count_subscribers(mPubConfMap->get_topic_name());
+      mDisparitySubnumber = count_subscribers(mPubDisparity->get_topic_name());
+    }
+  } catch (...) {
+    rcutils_reset_error();
+    RCLCPP_DEBUG(get_logger(), "publishImages: Exception while counting subscribers");
+    return 0;
+  }
+
+  return (mRgbSubnumber +
+         mRgbRawSubnumber +
+         mRgbGraySubnumber +
+         mRgbGrayRawSubnumber +
+         mLeftSubnumber +
+         mLeftRawSubnumber +
+         mLeftGraySubnumber +
+         mLeftGrayRawSubnumber +
+         mRightSubnumber +
+         mRightRawSubnumber +
+         mRightGraySubnumber +
+         mRightGrayRawSubnumber +
+         mStereoSubnumber +
+         mStereoRawSubnumber +
+         mDepthSubnumber +
+         mConfMapSubnumber +
+         mDisparitySubnumber +
+         mDepthInfoSubnumber) > 0;
+}
+
+void ZedCamera::retrieveVideoDepth()
+{
+  bool retrieved = false;
+  mRgbSubscribed = false;
+
+  // ----> Retrieve all required data
+  // RCLCPP_DEBUG(get_logger(), "Retrieving Video Data");
+  if (mRgbSubnumber + mLeftSubnumber + mStereoSubnumber > 0) {
+    retrieved |= sl::ERROR_CODE::SUCCESS ==
+      mZed.retrieveImage(mMatLeft, sl::VIEW::LEFT, sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatLeft.timestamp;
+    mRgbSubscribed = true;
+  }
+  if (mRgbRawSubnumber + mLeftRawSubnumber + mStereoRawSubnumber > 0) {
+    retrieved |=
+      sl::ERROR_CODE::SUCCESS ==
+      mZed.retrieveImage(mMatLeftRaw, sl::VIEW::LEFT_UNRECTIFIED, sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatLeftRaw.timestamp;
+  }
+  if (mRightSubnumber + mStereoSubnumber > 0) {
+    retrieved |= sl::ERROR_CODE::SUCCESS ==
+      mZed.retrieveImage(mMatRight, sl::VIEW::RIGHT, sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatRight.timestamp;
+  }
+  if (mRightRawSubnumber + mStereoRawSubnumber > 0) {
+    retrieved |= sl::ERROR_CODE::SUCCESS ==
+      mZed.retrieveImage(
+      mMatRightRaw, sl::VIEW::RIGHT_UNRECTIFIED, sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatRightRaw.timestamp;
+  }
+  if (mRgbGraySubnumber + mLeftGraySubnumber > 0) {
+    retrieved |=
+      sl::ERROR_CODE::SUCCESS ==
+      mZed.retrieveImage(mMatLeftGray, sl::VIEW::LEFT_GRAY, sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatLeftGray.timestamp;
+  }
+  if (mRgbGrayRawSubnumber + mLeftGrayRawSubnumber > 0) {
+    retrieved |= sl::ERROR_CODE::SUCCESS == mZed.retrieveImage(
+      mMatLeftRawGray, sl::VIEW::LEFT_UNRECTIFIED_GRAY,
+      sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatLeftRawGray.timestamp;
+  }
+  if (mRightGraySubnumber > 0) {
+    retrieved |=
+      sl::ERROR_CODE::SUCCESS ==
+      mZed.retrieveImage(mMatRightGray, sl::VIEW::RIGHT_GRAY, sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatRightGray.timestamp;
+  }
+  if (mRightGrayRawSubnumber > 0) {
+    retrieved |=
+      sl::ERROR_CODE::SUCCESS ==
+      mZed.retrieveImage(
+      mMatRightRawGray, sl::VIEW::RIGHT_UNRECTIFIED_GRAY, sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatRightRawGray.timestamp;
+  }
+  // RCLCPP_DEBUG(get_logger(), "Video Data retrieved");
+  // RCLCPP_DEBUG(get_logger(), "Retrieving Depth Data");
+  if (mDepthSubnumber > 0 || mDepthInfoSubnumber > 0) {
+    RCLCPP_DEBUG(get_logger(), "Retrieving Depth");
+    retrieved |=
+      sl::ERROR_CODE::SUCCESS ==
+      mZed.retrieveMeasure(mMatDepth, sl::MEASURE::DEPTH, sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatDepth.timestamp;
+  }
+  if (mDisparitySubnumber > 0) {
+    RCLCPP_DEBUG(get_logger(), "Retrieving Disparity");
+    retrieved |=
+      sl::ERROR_CODE::SUCCESS ==
+      mZed.retrieveMeasure(mMatDisp, sl::MEASURE::DISPARITY, sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatDisp.timestamp;
+  }
+  if (mConfMapSubnumber > 0) {
+    RCLCPP_DEBUG(get_logger(), "Retrieving Confidence");
+    retrieved |=
+      sl::ERROR_CODE::SUCCESS ==
+      mZed.retrieveMeasure(mMatConf, sl::MEASURE::CONFIDENCE, sl::MEM::CPU, mMatResol);
+    mGrabTS = mMatConf.timestamp;
+  }
+  if (mDepthInfoSubnumber > 0) {
+    retrieved |= sl::ERROR_CODE::SUCCESS == mZed.getCurrentMinMaxDepth(mMinDepth, mMaxDepth);
+#if ZED_SDK_MAJOR_VERSION == 3 && ZED_SDK_MINOR_VERSION == 7 && \
+    ZED_SDK_PATCH_VERSION == 0    // Units bug workaround
+    mMinDepth *= 0.001f;
+    mMaxDepth *= 0.001f;
+#endif
+    mGrabTS = mMatConf.timestamp;
+  }
+  // RCLCPP_DEBUG(get_logger(), "Depth Data retrieved");
+  // <---- Retrieve all required data
+}
+
+
+void ZedCamera::publishVideoDepth(rclcpp::Time & out_pub_ts)
 {
   sl_tools::StopWatch vdElabTimer;
 
   static sl::Timestamp lastZedTs = 0;  // Used to calculate stable publish frequency
 
-  bool rgb_subscribed = false;
+  // ----> Check RGB/Depth sync
+  static sl::Timestamp ts_rgb = 0;
+  static sl::Timestamp ts_depth = 0;
 
-  static size_t rgbSubnumber = 0;
-  static size_t rgbRawSubnumber = 0;
-  static size_t rgbGraySubnumber = 0;
-  static size_t rgbGrayRawSubnumber = 0;
-  static size_t leftSubnumber = 0;
-  static size_t leftRawSubnumber = 0;
-  static size_t leftGraySubnumber = 0;
-  static size_t leftGrayRawSubnumber = 0;
-  static size_t rightSubnumber = 0;
-  static size_t rightRawSubnumber = 0;
-  static size_t rightGraySubnumber = 0;
-  static size_t rightGrayRawSubnumber = 0;
-  static size_t stereoSubnumber = 0;
-  static size_t stereoRawSubnumber = 0;
-  static size_t depthSubnumber = 0;
-  static size_t confMapSubnumber = 0;
-  static size_t disparitySubnumber = 0;
-  static size_t depthInfoSubnumber = 0;
+  if (mRgbSubscribed && (mDepthSubnumber > 0 || mDepthInfoSubnumber > 0)) {
+    ts_rgb = mMatLeft.timestamp;
+    ts_depth = mMatDepth.timestamp;
 
-  bool retrieved = false;
-
-  static sl::Mat mat_left, mat_left_raw;
-  static sl::Mat mat_right, mat_right_raw;
-  static sl::Mat mat_left_gray, mat_left_raw_gray;
-  static sl::Mat mat_right_gray, mat_right_raw_gray;
-  static sl::Mat mat_depth, mat_disp, mat_conf;
-
-  static sl::Timestamp ts_rgb = 0;    // used to check RGB/Depth sync
-  static sl::Timestamp ts_depth = 0;  // used to check RGB/Depth sync
-  static sl::Timestamp grab_ts = 0;
-
-  float min_depth = 0.0f, max_depth = 0.0f;
-
-  try {
-    rgbSubnumber = mPubRgb.getNumSubscribers();
-    rgbRawSubnumber = mPubRawRgb.getNumSubscribers();
-    rgbGraySubnumber = mPubRgbGray.getNumSubscribers();
-    rgbGrayRawSubnumber = mPubRawRgbGray.getNumSubscribers();
-    leftSubnumber = mPubLeft.getNumSubscribers();
-    leftRawSubnumber = mPubRawLeft.getNumSubscribers();
-    leftGraySubnumber = mPubLeftGray.getNumSubscribers();
-    leftGrayRawSubnumber = mPubRawLeftGray.getNumSubscribers();
-    rightSubnumber = mPubRight.getNumSubscribers();
-    rightRawSubnumber = mPubRawRight.getNumSubscribers();
-    rightGraySubnumber = mPubRightGray.getNumSubscribers();
-    rightGrayRawSubnumber = mPubRawRightGray.getNumSubscribers();
-    stereoSubnumber = mPubStereo.getNumSubscribers();
-    stereoRawSubnumber = mPubRawStereo.getNumSubscribers();
-
-    if (!mDepthDisabled) {
-      depthSubnumber = mPubDepth.getNumSubscribers();
-      depthInfoSubnumber = count_subscribers(mPubDepthInfo->get_topic_name());
-      confMapSubnumber = count_subscribers(mPubConfMap->get_topic_name());
-      disparitySubnumber = count_subscribers(mPubDisparity->get_topic_name());
-    }
-  } catch (...) {
-    rcutils_reset_error();
-    RCLCPP_DEBUG(get_logger(), "publishImages: Exception while counting subscribers");
-    return false;
-  }
-
-  // ----> Wait for data grabbed
-  while (!mZedDataReady) {
-    // Check thread stopping
-    if (mThreadStop) {
-      return false;
-    } else {
-      rclcpp::sleep_for(500us);
-      continue;
-    }
-  }
-  mZedDataReady = false;
-  std::unique_lock<std::mutex> datalock(mCamDataMutex);
-  // <---- Wait for data grabbed
-
-  // Start processing timer for diagnostic
-  vdElabTimer.tic();
-
-  // ----> Retrieve all required data
-  // RCLCPP_DEBUG(get_logger(), "Retrieving Video Data");
-  if (rgbSubnumber + leftSubnumber + stereoSubnumber > 0) {
-    retrieved |= sl::ERROR_CODE::SUCCESS ==
-      mZed.retrieveImage(mat_left, sl::VIEW::LEFT, sl::MEM::CPU, mMatResol);
-    ts_rgb = mat_left.timestamp;
-    grab_ts = mat_left.timestamp;
-    rgb_subscribed = true;
-  }
-  if (rgbRawSubnumber + leftRawSubnumber + stereoRawSubnumber > 0) {
-    retrieved |=
-      sl::ERROR_CODE::SUCCESS ==
-      mZed.retrieveImage(mat_left_raw, sl::VIEW::LEFT_UNRECTIFIED, sl::MEM::CPU, mMatResol);
-    grab_ts = mat_left_raw.timestamp;
-  }
-  if (rightSubnumber + stereoSubnumber > 0) {
-    retrieved |= sl::ERROR_CODE::SUCCESS ==
-      mZed.retrieveImage(mat_right, sl::VIEW::RIGHT, sl::MEM::CPU, mMatResol);
-    grab_ts = mat_right.timestamp;
-  }
-  if (rightRawSubnumber + stereoRawSubnumber > 0) {
-    retrieved |= sl::ERROR_CODE::SUCCESS ==
-      mZed.retrieveImage(
-      mat_right_raw, sl::VIEW::RIGHT_UNRECTIFIED, sl::MEM::CPU, mMatResol);
-    grab_ts = mat_right_raw.timestamp;
-  }
-  if (rgbGraySubnumber + leftGraySubnumber > 0) {
-    retrieved |=
-      sl::ERROR_CODE::SUCCESS ==
-      mZed.retrieveImage(mat_left_gray, sl::VIEW::LEFT_GRAY, sl::MEM::CPU, mMatResol);
-    grab_ts = mat_left_gray.timestamp;
-  }
-  if (rgbGrayRawSubnumber + leftGrayRawSubnumber > 0) {
-    retrieved |= sl::ERROR_CODE::SUCCESS == mZed.retrieveImage(
-      mat_left_raw_gray, sl::VIEW::LEFT_UNRECTIFIED_GRAY,
-      sl::MEM::CPU, mMatResol);
-    grab_ts = mat_left_raw_gray.timestamp;
-  }
-  if (rightGraySubnumber > 0) {
-    retrieved |=
-      sl::ERROR_CODE::SUCCESS ==
-      mZed.retrieveImage(mat_right_gray, sl::VIEW::RIGHT_GRAY, sl::MEM::CPU, mMatResol);
-    grab_ts = mat_right_gray.timestamp;
-  }
-  if (rightGrayRawSubnumber > 0) {
-    retrieved |=
-      sl::ERROR_CODE::SUCCESS ==
-      mZed.retrieveImage(
-      mat_right_raw_gray, sl::VIEW::RIGHT_UNRECTIFIED_GRAY, sl::MEM::CPU, mMatResol);
-    grab_ts = mat_right_raw_gray.timestamp;
-  }
-  // RCLCPP_DEBUG(get_logger(), "Video Data retrieved");
-  // RCLCPP_DEBUG(get_logger(), "Retrieving Depth Data");
-  if (depthSubnumber > 0 || depthInfoSubnumber > 0) {
-    RCLCPP_DEBUG(get_logger(), "Retrieving Depth");
-    retrieved |=
-      sl::ERROR_CODE::SUCCESS ==
-      mZed.retrieveMeasure(mat_depth, sl::MEASURE::DEPTH, sl::MEM::CPU, mMatResol);
-    grab_ts = mat_depth.timestamp;
-
-    ts_depth = mat_depth.timestamp;
-
-    if (rgb_subscribed && (ts_rgb.data_ns != 0 && (ts_depth.data_ns != ts_rgb.data_ns))) {
+    if (mRgbSubscribed && (ts_rgb.data_ns != 0 && (ts_depth.data_ns != ts_rgb.data_ns))) {
       RCLCPP_WARN_STREAM(
         get_logger(), "!!!!! DEPTH/RGB ASYNC !!!!! - Delta: "
           << 1e-9 * static_cast<double>(ts_depth - ts_rgb) << " sec");
     }
   }
-  if (disparitySubnumber > 0) {
-    RCLCPP_DEBUG(get_logger(), "Retrieving Disparity");
-    retrieved |=
-      sl::ERROR_CODE::SUCCESS ==
-      mZed.retrieveMeasure(mat_disp, sl::MEASURE::DISPARITY, sl::MEM::CPU, mMatResol);
-    grab_ts = mat_disp.timestamp;
-  }
-  if (confMapSubnumber > 0) {
-    RCLCPP_DEBUG(get_logger(), "Retrieving Confidence");
-    retrieved |=
-      sl::ERROR_CODE::SUCCESS ==
-      mZed.retrieveMeasure(mat_conf, sl::MEASURE::CONFIDENCE, sl::MEM::CPU, mMatResol);
-    grab_ts = mat_conf.timestamp;
-  }
-  if (depthInfoSubnumber > 0) {
-    retrieved |= sl::ERROR_CODE::SUCCESS == mZed.getCurrentMinMaxDepth(min_depth, max_depth);
-#if ZED_SDK_MAJOR_VERSION == 3 && ZED_SDK_MINOR_VERSION == 7 && \
-    ZED_SDK_PATCH_VERSION == 0    // Units bug workaround
-    min_depth *= 0.001f;
-    max_depth *= 0.001f;
-#endif
-  }
-  // RCLCPP_DEBUG(get_logger(), "Depth Data retrieved");
-  // <---- Retrieve all required data
+  // <---- Check RGB/Depth sync
 
-  // Data retrieved, we can release the mutex
-  datalock.unlock();
-
-  if (!retrieved) {
-    lastZedTs = 0;
-    out_pub_ts = TIMEZERO_ROS;
-    return false;
-  }
+  // Start processing timer for diagnostic
+  vdElabTimer.tic();
 
   // ----> Check if a grab has been done before publishing the same images
-  if (grab_ts.data_ns == lastZedTs.data_ns) {
+  if (mGrabTS.data_ns == lastZedTs.data_ns) {
     out_pub_ts = TIMEZERO_ROS;
     // Data not updated by a grab calling in the grab thread
     RCLCPP_DEBUG_STREAM(get_logger(), "publishVideoDepth: ignoring not update data");
-    return true;
+    return;
   }
 
   if (lastZedTs.data_ns != 0) {
-    double period_sec = static_cast<double>(grab_ts.data_ns - lastZedTs.data_ns) / 1e9;
+    double period_sec = static_cast<double>(mGrabTS.data_ns - lastZedTs.data_ns) / 1e9;
     RCLCPP_DEBUG_STREAM(
       get_logger(),
       "VIDEO/DEPTH PUB LAST PERIOD: " << period_sec << " sec @" << 1. / period_sec << " Hz");
@@ -4436,12 +4457,12 @@ bool ZedCamera::publishVideoDepth(rclcpp::Time & out_pub_ts)
       "VIDEO/DEPTH PUB MEAN PERIOD: " << mVideoDepthPeriodMean_sec->getAvg() << " sec @"
                                       << 1. / mVideoDepthPeriodMean_sec->getAvg() << " Hz");
   }
-  lastZedTs = grab_ts;
+  lastZedTs = mGrabTS;
   // <---- Check if a grab has been done before publishing the same images
 
   static rclcpp::Time timeStamp;
   if (!mSvoMode) {
-    timeStamp = sl_tools::slTime2Ros(grab_ts, get_clock()->get_clock_type());
+    timeStamp = sl_tools::slTime2Ros(mGrabTS, get_clock()->get_clock_type());
   } else {
     // timeStamp = sl_tools::slTime2Ros(mZed.getTimestamp(sl::TIME_REFERENCE::CURRENT),
     // get_clock()->get_clock_type());
@@ -4451,117 +4472,117 @@ bool ZedCamera::publishVideoDepth(rclcpp::Time & out_pub_ts)
   out_pub_ts = timeStamp;
 
   // ----> Publish the left=rgb image if someone has subscribed to
-  if (leftSubnumber > 0) {
-    publishImageWithInfo(mat_left, mPubLeft, mLeftCamInfoMsg, mLeftCamOptFrameId, timeStamp);
+  if (mLeftSubnumber > 0) {
+    publishImageWithInfo(mMatLeft, mPubLeft, mLeftCamInfoMsg, mLeftCamOptFrameId, timeStamp);
   }
-  if (rgbSubnumber > 0) {
-    publishImageWithInfo(mat_left, mPubRgb, mRgbCamInfoMsg, mDepthOptFrameId, timeStamp);
+  if (mRgbSubnumber > 0) {
+    publishImageWithInfo(mMatLeft, mPubRgb, mRgbCamInfoMsg, mDepthOptFrameId, timeStamp);
   }
   // <---- Publish the left=rgb image if someone has subscribed to
 
   // ----> Publish the left_raw=rgb_raw image if someone has subscribed to
-  if (leftRawSubnumber > 0) {
+  if (mLeftRawSubnumber > 0) {
     publishImageWithInfo(
-      mat_left_raw, mPubRawLeft, mLeftCamInfoRawMsg, mLeftCamOptFrameId, timeStamp);
+      mMatLeftRaw, mPubRawLeft, mLeftCamInfoRawMsg, mLeftCamOptFrameId, timeStamp);
   }
-  if (rgbRawSubnumber > 0) {
-    publishImageWithInfo(mat_left_raw, mPubRawRgb, mRgbCamInfoRawMsg, mDepthOptFrameId, timeStamp);
+  if (mRgbRawSubnumber > 0) {
+    publishImageWithInfo(mMatLeftRaw, mPubRawRgb, mRgbCamInfoRawMsg, mDepthOptFrameId, timeStamp);
   }
   // <---- Publish the left_raw=rgb_raw image if someone has subscribed to
 
   // ----> Publish the left_gray=rgb_gray image if someone has subscribed to
-  if (leftGraySubnumber > 0) {
+  if (mLeftGraySubnumber > 0) {
     publishImageWithInfo(
-      mat_left_gray, mPubLeftGray, mLeftCamInfoMsg, mLeftCamOptFrameId, timeStamp);
+      mMatLeftGray, mPubLeftGray, mLeftCamInfoMsg, mLeftCamOptFrameId, timeStamp);
   }
-  if (rgbGraySubnumber > 0) {
-    publishImageWithInfo(mat_left_gray, mPubRgbGray, mRgbCamInfoMsg, mDepthOptFrameId, timeStamp);
+  if (mRgbGraySubnumber > 0) {
+    publishImageWithInfo(mMatLeftGray, mPubRgbGray, mRgbCamInfoMsg, mDepthOptFrameId, timeStamp);
   }
   // <---- Publish the left_raw=rgb_raw image if someone has subscribed to
 
   // ----> Publish the left_raw_gray=rgb_raw_gray image if someone has
   // subscribed to
-  if (leftGrayRawSubnumber > 0) {
+  if (mLeftGrayRawSubnumber > 0) {
     publishImageWithInfo(
-      mat_left_raw_gray, mPubRawLeftGray, mLeftCamInfoRawMsg, mLeftCamOptFrameId, timeStamp);
+      mMatLeftRawGray, mPubRawLeftGray, mLeftCamInfoRawMsg, mLeftCamOptFrameId, timeStamp);
   }
-  if (rgbGrayRawSubnumber > 0) {
+  if (mRgbGrayRawSubnumber > 0) {
     publishImageWithInfo(
-      mat_left_raw_gray, mPubRawRgbGray, mRgbCamInfoRawMsg, mDepthOptFrameId, timeStamp);
+      mMatLeftRawGray, mPubRawRgbGray, mRgbCamInfoRawMsg, mDepthOptFrameId, timeStamp);
   }
   // ----> Publish the left_raw_gray=rgb_raw_gray image if someone has
   // subscribed to
 
   // ----> Publish the right image if someone has subscribed to
-  if (rightSubnumber > 0) {
-    publishImageWithInfo(mat_right, mPubRight, mRightCamInfoMsg, mRightCamOptFrameId, timeStamp);
+  if (mRightSubnumber > 0) {
+    publishImageWithInfo(mMatRight, mPubRight, mRightCamInfoMsg, mRightCamOptFrameId, timeStamp);
   }
   // <---- Publish the right image if someone has subscribed to
 
   // ----> Publish the right raw image if someone has subscribed to
-  if (rightRawSubnumber > 0) {
+  if (mRightRawSubnumber > 0) {
     publishImageWithInfo(
-      mat_right_raw, mPubRawRight, mRightCamInfoRawMsg, mRightCamOptFrameId, timeStamp);
+      mMatRightRaw, mPubRawRight, mRightCamInfoRawMsg, mRightCamOptFrameId, timeStamp);
   }
   // <---- Publish the right raw image if someone has subscribed to
 
   // ----> Publish the right gray image if someone has subscribed to
-  if (rightGraySubnumber > 0) {
+  if (mRightGraySubnumber > 0) {
     publishImageWithInfo(
-      mat_right_gray, mPubRightGray, mRightCamInfoMsg, mRightCamOptFrameId, timeStamp);
+      mMatRightGray, mPubRightGray, mRightCamInfoMsg, mRightCamOptFrameId, timeStamp);
   }
   // <---- Publish the right gray image if someone has subscribed to
 
   // ----> Publish the right raw gray image if someone has subscribed to
-  if (rightGrayRawSubnumber > 0) {
+  if (mRightGrayRawSubnumber > 0) {
     publishImageWithInfo(
-      mat_right_raw_gray, mPubRawRightGray, mRightCamInfoRawMsg, mRightCamOptFrameId, timeStamp);
+      mMatRightRawGray, mPubRawRightGray, mRightCamInfoRawMsg, mRightCamOptFrameId, timeStamp);
   }
   // <---- Publish the right raw gray image if someone has subscribed to
 
   // ----> Publish the side-by-side image if someone has subscribed to
-  if (stereoSubnumber > 0) {
-    auto combined = sl_tools::imagesToROSmsg(mat_left, mat_right, mCameraFrameId, timeStamp);
+  if (mStereoSubnumber > 0) {
+    auto combined = sl_tools::imagesToROSmsg(mMatLeft, mMatRight, mCameraFrameId, timeStamp);
     RCLCPP_DEBUG_STREAM(get_logger(), "Publishing SIDE-BY-SIDE message");
     mPubStereo.publish(combined);
   }
   // <---- Publish the side-by-side image if someone has subscribed to
 
   // ----> Publish the side-by-side image if someone has subscribed to
-  if (stereoRawSubnumber > 0) {
+  if (mStereoRawSubnumber > 0) {
     auto combined =
-      sl_tools::imagesToROSmsg(mat_left_raw, mat_right_raw, mCameraFrameId, timeStamp);
+      sl_tools::imagesToROSmsg(mMatLeftRaw, mMatRightRaw, mCameraFrameId, timeStamp);
     RCLCPP_DEBUG_STREAM(get_logger(), "Publishing SIDE-BY-SIDE RAW message");
     mPubRawStereo.publish(combined);
   }
   // <---- Publish the side-by-side image if someone has subscribed to
 
   // ---->  Publish the depth image if someone has subscribed to
-  if (depthSubnumber > 0) {
-    publishDepthMapWithInfo(mat_depth, timeStamp);
+  if (mDepthSubnumber > 0) {
+    publishDepthMapWithInfo(mMatDepth, timeStamp);
   }
   // <----  Publish the depth image if someone has subscribed to
 
   // ---->  Publish the confidence image and map if someone has subscribed to
-  if (confMapSubnumber > 0) {
+  if (mConfMapSubnumber > 0) {
     RCLCPP_DEBUG_STREAM(get_logger(), "Publishing CONF MAP message");
-    mPubConfMap->publish(*sl_tools::imageToROSmsg(mat_conf, mDepthOptFrameId, timeStamp));
+    mPubConfMap->publish(*sl_tools::imageToROSmsg(mMatConf, mDepthOptFrameId, timeStamp));
   }
   // <----  Publish the confidence image and map if someone has subscribed to
 
   // ----> Publish the disparity image if someone has subscribed to
-  if (disparitySubnumber > 0) {
-    publishDisparity(mat_disp, timeStamp);
+  if (mDisparitySubnumber > 0) {
+    publishDisparity(mMatDisp, timeStamp);
   }
   // <---- Publish the disparity image if someone has subscribed to
 
   // ----> Publish the depth info if someone has subscribed to
-  if (depthInfoSubnumber > 0) {
+  if (mDepthInfoSubnumber > 0) {
     depthInfoMsgPtr depthInfoMsg = std::make_unique<zed_interfaces::msg::DepthInfoStamped>();
     depthInfoMsg->header.stamp = timeStamp;
     depthInfoMsg->header.frame_id = mDepthOptFrameId;
-    depthInfoMsg->min_depth = min_depth;
-    depthInfoMsg->max_depth = max_depth;
+    depthInfoMsg->min_depth = mMinDepth;
+    depthInfoMsg->max_depth = mMaxDepth;
 
     RCLCPP_DEBUG_STREAM(get_logger(), "Publishing DEPTH INFO message");
     mPubDepthInfo->publish(std::move(depthInfoMsg));
@@ -4570,8 +4591,6 @@ bool ZedCamera::publishVideoDepth(rclcpp::Time & out_pub_ts)
 
   // Diagnostic statistic
   mVideoDepthElabMean_sec->addValue(vdElabTimer.toc());
-
-  return true;
 }
 
 void ZedCamera::publishImageWithInfo(
@@ -5973,19 +5992,31 @@ void ZedCamera::callback_updateDiagnostic(diagnostic_updater::DiagnosticStatusWr
   }
 
   if (mGrabStatus == sl::ERROR_CODE::SUCCESS) {
-    stat.summary(diagnostic_msgs::msg::DiagnosticStatus::OK, "Camera grabbing");
-
     double freq = 1. / mGrabPeriodMean_sec->getAvg();
     double freq_perc = 100. * freq / mCamGrabFrameRate;
     stat.addf("Capture", "Mean Frequency: %.1f Hz (%.1f%%)", freq, freq_perc);
-    stat.addf(
-      "Capture", "SDK Grab duration: %.3f sec (Max. %.3f sec)", mGrabFuncMean_sec->getAvg(),
-      1. / mCamGrabFrameRate);
-    stat.addf(
-      "Capture", "Tot. Processing Time: %.6f sec (Max. %.3f sec)", mElabPeriodMean_sec->getAvg(),
-      1. / mCamGrabFrameRate);
 
-    if (mPublishingData) {
+    double frame_proc_sec = mElabPeriodMean_sec->getAvg();
+    double frame_grab_period = 1. / mCamGrabFrameRate;
+    stat.addf(
+      "Capture", "Tot. Processing Time: %.6f sec (Max. %.3f sec)", frame_proc_sec,
+      frame_grab_period);
+
+    static int overload_count = 0;
+    if (frame_proc_sec > frame_grab_period) {
+      overload_count++;
+    }
+
+    if (overload_count >= 10) {
+      stat.summary(
+        diagnostic_msgs::msg::DiagnosticStatus::WARN,
+        "System overloaded. 'grab_frame_rate' too high.");
+    } else {
+      overload_count = 0;
+      stat.summary(diagnostic_msgs::msg::DiagnosticStatus::OK, "Camera grabbing");
+    }
+
+    if (mVdPublishing) {
       freq = 1. / mVideoDepthPeriodMean_sec->getAvg();
       freq_perc = 100. * freq / mPubFrameRate;
       stat.addf("Video/Depth", "Mean Frequency: %.1f Hz (%.1f%%)", freq, freq_perc);